+++ Assets/CommonCore/CoreShared/UI/LoadGamePanelController.cs
@@ -163,9 +163,9 @@ namespace CommonCore.UI
 
             var metadata = SaveUtils.LoadSaveMetadata(Saves[SelectedSaveIndex].FileName);
 
-            if(metadata == null || metadata.GameVersion.GameVersion < CoreParams.GameVersion)
+            if(metadata == null || metadata?.GameVersion?.GameVersion == null || metadata.GameVersion.GameVersion < CoreParams.GameVersion)
             {
-                string message = string.Format(Sub.Replace("SaveVersionMismatchMessage", "IGUI_SAVE"), metadata?.GameVersion?.ToString() ?? "unknown", CoreParams.GameVersion);
+                string message = string.Format(Sub.Replace("SaveVersionMismatchMessage", "IGUI_SAVE"), metadata?.GameVersion?.GameVersion?.ToString() ?? "unknown", CoreParams.GameVersion);
                 Modal.PushConfirmModal(message, Sub.Replace("SaveVersionMismatch", "IGUI_SAVE"), "Load", "Cancel", "", (status, tag, result) =>
                 {
                     if(status == ModalStatusCode.Complete && result)